2001-04-04  Patrick Eriksson  <patrick@rss.chalmers.se>
	* arts-0-0-152

	* NaNs were obtained when using float and a conversion to brightness
	temperatures was performed. Some scalars have been hardcoded to be
	double to avoid this problem.

	* A similar problem was found in the vector version of planck.

	* A warning, when using very large or small values, as the Planck and
	Boltzman constants, I would recomment to hard-code double to avoid
	these disturbing problems.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@844 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/atm_funcs.cc
@@ -77,8 +77,8 @@ void planck (
         const VECTOR&     t )
 {
   // Double must be used here (if not, a becomes 0 when using float)
-  static const double a = 2.0*PLANCK_CONST/(SPEED_OF_LIGHT*SPEED_OF_LIGHT);
-  static const double b = PLANCK_CONST/BOLTZMAN_CONST;
+  static const double  a = 2.0*PLANCK_CONST/(SPEED_OF_LIGHT*SPEED_OF_LIGHT);
+  static const double  b = PLANCK_CONST/BOLTZMAN_CONST;
 
   const size_t    n_f  = f.size();
   const size_t    n_t  = t.size();
@@ -133,9 +133,10 @@ void planck (
        const VECTOR&    f,
        const Numeric&   t )
 {
-  static const Numeric a = 2.0*PLANCK_CONST/(SPEED_OF_LIGHT*SPEED_OF_LIGHT);
-  static const Numeric b = PLANCK_CONST/BOLTZMAN_CONST;
-         const Numeric c = b/t; 
+  // Double must be used here (if not, a becomes 0 when using float)
+  static const double  a = 2.0*PLANCK_CONST/(SPEED_OF_LIGHT*SPEED_OF_LIGHT);
+  static const double  b = PLANCK_CONST/BOLTZMAN_CONST;
+         const double  c = b/t; 
 
   assert( B.size()==f.size() );
